openingd: pull out check_configs into a common place
We'll re-use it for dualopend!
