+++ common/initial_commit_tx.h
@@ -10,6 +10,7 @@
 
 struct bitcoin_txid;
 struct keyset;
+struct wally_tx_output;
 
 /* BOLT #3:
  *
